make `anvil` easier accessible as a library
